NO-JIRA Move files to Src directory.
